refactor: move tests
